Move UploadFile to commonMain
